Ahora si

+++ OmniMonitor/Server/Controllers/DatasetAMController.cs
@@ -156,9 +156,12 @@ namespace OmniMonitor.Server.Controllers
                     return NotFound($"No se encontró el DatasetAM con ID {id} para el usuario {request.Username}.");
                 }
 
+                // Primero validar el nombre en la tabla general antes de actualizar cualquier tabla
+                await _datasetUMService.ValidateDatasetNameAsync(request.Nombre, request.Username, ModuleType.AssetManager, existingDataset.DatasetId);
+
                 // Llamar al servicio que incluye la validación de nombres únicos
                 DatasetAM updatedDataset = await _datasetAmService.UpdateDatasetAMAsync(existingDataset, request);
-                var requestDataset = new CreateDatasetRequest(existingDataset.Nombre, request.Username, ModuleType.AssetManager);
+                var requestDataset = new CreateDatasetRequest(request.Nombre, request.Username, ModuleType.AssetManager);
                 Datasets newDataset = await _datasetUMService.UpdateDatasetAsyncAM(updatedDataset.DatasetId, requestDataset, updatedDataset);
                 return Ok(updatedDataset);
             }
@@ -191,8 +194,14 @@ namespace OmniMonitor.Server.Controllers
                 string username = await _sondaAuthService.GetUserByTokenOMAsync(token);
                 DatasetAM? datasetid = await _context.DatasetAM
                 .FirstOrDefaultAsync(d => d.Id_Dataset == id && d.Username == username);
+                
+                if (datasetid == null)
+                {
+                    return NotFound($"No se encontró el dataset con ID {id} para el usuario {username}.");
+                }
+                
                 await _datasetAmService.DeleteDatasetAMAsync(id, username);
-                await _datasetUMService.DeleteDatasetAsync(datasetid!.DatasetId, username);
+                await _datasetUMService.DeleteDatasetAsync(datasetid.DatasetId, username);
                 return NoContent();
             }
             catch (Exception ex)
